rename _trajectory package to trajectory
