+++ osmosis_ai/rollout/trajectory/__init__.py
@@ -1,11 +1,11 @@
 """Save finished rollouts as ATIF trajectory documents (backend-agnostic)."""
 
-from osmosis_ai.rollout._trajectory.converter import convert_sample_to_trajectory
-from osmosis_ai.rollout._trajectory.report import (
+from osmosis_ai.rollout.trajectory.converter import convert_sample_to_trajectory
+from osmosis_ai.rollout.trajectory.report import (
     TrajectoryReport,
     report_from_response,
 )
-from osmosis_ai.rollout._trajectory.save import save_trajectories
+from osmosis_ai.rollout.trajectory.save import save_trajectories
 
 __all__ = [
     "TrajectoryReport",
